refactor(engine): Task 5 — share stroke-play standings live + on Finish
Relocate the pure tie/leader logic to a client-safe shared module
(src/lib/strokePlay.ts, no server/DB deps) so it's the ONE implementation used
by both:
 (a) the live standings strip — client sums the loaded score_entries for running
     totals + 'Leading' (position === 1) as scores land, no per-keystroke write;
 (b) the persisted final record — server computeStrokePlayResults writes
     game_results on Finish (idempotent), importing the same pure function.

The lowest-total-leads / ties-share-position rule can no longer diverge between
the live strip and the Final screen. tsc/lint/pure test green.
