+++ src/lib/strokePlay.test.ts
